+++ packages/react-storefront-moov-xdn/src/getHeaders.js
