+++ packages/react-storefront/src/router/index.js
@@ -8,4 +8,3 @@ export { default as fromClient } from './fromClient'
 export { default as proxyUpstream } from './proxyUpstream'
 export { default as cache } from './cache'
 export { default as withGlobalState } from './withGlobalState'
-export { default as Response } from './Response'
